% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance.R
\name{balance}
\alias{balance}
\title{This is the function to calculate balance; allowing NA values in z corresponds to undetermined group status}
\usage{
balance(X, z)
}
\arguments{
\item{X}{is the relative abundance matrix}

\item{z}{is the group indicator vector;}
}
\description{
This is the function to calculate balance; allowing NA values in z corresponds to undetermined group status
}
