% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BalReg_mcmc.R
\name{BalReg_mcmc}
\alias{BalReg_mcmc}
\title{This is the main function to perform MCMC sampling for Balance Regression in linear outcomes}
\usage{
BalReg_mcmc(beta0, V0, lambda, v, w1, w2, Y, X, verbose, iter, start)
}
\arguments{
\item{beta0}{mean of a multivariate normal distribution which is the prior distribution for regression coefficients}

\item{V0}{diagnol elements of the variance-covariance matrix of a multivariate normal distribution which is the prior distribution for regression coefficients}

\item{lambda, v}{parameters of the inverse gamma distribution which is the prior distribution of model variance}

\item{w1, w2}{parameter of prior distribution for the balance configuration vector z}

\item{Y}{outcome vector}

\item{X}{compositional matrix}

\item{verbose}{boolean; indicating whether the program will output verbose progress report}

\item{iter}{total number of iteration to run}

\item{start}{starting value of Balance configuration vector z}
}
\description{
This is the main function to perform MCMC sampling for Balance Regression in linear outcomes
}
