% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BalReg_probit_mcmc.R
\name{BalReg_probit_mcmc}
\alias{BalReg_probit_mcmc}
\title{This is the MCMC set up for Balance Regression with binary outcome. The three major steps are:
   initialize with z
   sample y*|z,y
   update z}
\usage{
BalReg_probit_mcmc(beta0, V0, w1, w2, Y, X, verbose, iter, start,
  filename = "log.txt")
}
\arguments{
\item{beta0}{mean of a multivariate normal distribution which is the prior distribution for regression coefficients}

\item{V0}{diagnol elements of the variance-covariance matrix of a multivariate normal distribution which is the prior distribution for regression coefficients}

\item{w1, w2}{parameter of prior distribution for the balance configuration vector z}

\item{Y}{outcome vector}

\item{X}{compositional matrix}

\item{verbose}{boolean; indicating whether the program will output verbose progress report}

\item{iter}{total number of iteration to run}

\item{start}{starting value of Balance configuration vector z}

\item{filename}{name of the file to output verbose progress, default to log.txt}
}
\description{
This is the MCMC set up for Balance Regression with binary outcome. The three major steps are:
   initialize with z
   sample y*|z,y
   update z
}
