% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logtarget.R
\name{logtarget}
\alias{logtarget}
\title{calculating the log target value for the target function.}
\usage{
logtarget(z, a0, h, b0, varc, lambda, v, w1, w2, Y, X)
}
\description{
calculating the log target value for the target function.
}
