% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh_update.R
\name{mh_update}
\alias{mh_update}
\title{Metroplis Hastings Algorithm
proposal density:
 phi0 probability for changes between 0 and 1 (addition,deletion and switch)
 phi1 probability for changes between 0 and -1
 phi2 probaiblity for changes between 1 and -1
 pa, probability for addition
 pd,probability for deletion
 pw, probability for swtich}
\usage{
mh_update(zt, phi0, phi1, pa, pd, pw, verbose, ...)
}
\description{
Metroplis Hastings Algorithm
proposal density:
 phi0 probability for changes between 0 and 1 (addition,deletion and switch)
 phi1 probability for changes between 0 and -1
 phi2 probaiblity for changes between 1 and -1
 pa, probability for addition
 pd,probability for deletion
 pw, probability for swtich
}
